refactor: 타입 도메인별 분리 및 index.ts로 export
