+++ src/api/fetchRandomStoreList.tsx
@@ -1,4 +1,4 @@
-import { CampusId, Store } from "types/commonTypes";
+import { CampusId, Store } from "types/common";
 
 import { ENDPOINTS } from "constants/api";
