+++ src/components/pages/CampusSelectPage/CampusSelectPage.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import { Campus } from "types/commonTypes";
+import { Campus } from "types/common";
 
 import CAMPUS from "constants/campus";
 import { PATHNAME } from "constants/routes";
