+++ src/components/pages/SearchResultPage/SearchResultPage.tsx
@@ -3,7 +3,7 @@ import { useContext } from "react";
 import { MdArrowBackIos } from "react-icons/md";
 import { useInfiniteQuery } from "react-query";
 import { useNavigate, useSearchParams } from "react-router-dom";
-import { Campus, Store } from "types/commonTypes";
+import { Campus, Store } from "types/common";
 
 import { NETWORK, SIZE } from "constants/api";
 import { getCampusId } from "constants/campus";
