+++ src/types/common/authorTypes.ts
@@ -1,0 +1,4 @@
+export interface Author {
+  username: string;
+  profileImage: string;
+}
