+++ src/types/common/campusTypes.ts
@@ -1,0 +1,2 @@
+export type CampusId = 1 | 2;
+export type Campus = "잠실" | "선릉";
