+++ src/types/common/categoryTypes.ts
@@ -1,0 +1,1 @@
+export type CategoryId = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8";
