+++ src/types/common/index.ts
@@ -1,0 +1,5 @@
+export * from "./authorTypes";
+export * from "./campusTypes";
+export * from "./categoryTypes";
+export * from "./reviewTypes";
+export * from "./storeTypes";
